phase_2/pyproject.toml
[project]
name = "phase-2-react-agent"
version = "0.1.0"
description = "Learn the ReAct (Reasoning + Acting) pattern"
requires-python = ">=3.11"
dependencies = [
    "python-dotenv>=1.2.2",
    "agent-core",
]

[tool.uv.sources]
agent-core = { path = "../lib", editable = true }

[tool.ruff]
line-length = 100
